# cla-backend-go/swagger/cla.v2.yaml
...
swagger: '2.0'
info:
  title: EasyCLA v2 API
  version: '2.0'
  x-logo:
    url: "https://www.linuxfoundation.org/wp-content/uploads/2017/06/LF_logo_NEW_H_W.png"
    backgroundColor: "#FFFFFF"
    altText: "The Linux Foundation - EasyCLA API"
basePath: /v4
securityDefinitions:
  lf-auth:
    type: apiKey
    name: X-ACL
    in: header
    description: Requires X-ACL headers and bearer token
security:
  - lf-auth: [ ]
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /gitlab/oauth/callback:
    get:
      summary: The endpoint is called after user authorizes EasyCLA bot
      description: The endpoint is responsible for storing the access token for the user and registering the webhooks is autoenable is on
      security: []
      operationId: gitlabOauthCallback
      parameters:
        - name: code
          description: oauth code used to fetch the access token
          in: query
          type: string
          required: true
        - name: state
          description: state is used to find the gitlab organization
          in: query
          type: string
          required: true
      responses:
        '200':
          description: 'Success'
          headers:
            x-request-id:
              type: string
              description: The unique request ID value - assigned/set by the API Gateway based on the session
          schema:
            $ref: '#/definitions/success-response'
        '400':
          $ref: '#/responses/invalid-request'
        '403':
          $ref: '#/responses/forbidden'
        '404':
          $ref: '#/responses/not-found'
        '500':
          $ref: '#/responses/internal-server-error'
      tags:
        - gitlab-activity
responses:
...
  invalid-request:
    description: Invalid request
    headers:
      x-request-id:
        type: string
        description: The unique request ID value - assigned/set by the API Gateway based on the session
    schema:
      $ref: '#/definitions/error-response'
  forbidden:
    description: Insufficient privilege to execute action
    headers:
      x-request-id:
        type: string
        description: The unique request ID value - assigned/set by the API Gateway based on the session
    schema:
      $ref: '#/definitions/error-response'
  not-found:
    description: Not found
    headers:
      x-request-id:
        type: string
        description: The unique request ID value - assigned/set by the API Gateway based on the session
    schema:
      $ref: '#/definitions/error-response'
  internal-server-error:
    description: Internal server error
    headers:
      x-request-id:
        type: string
        description: The unique request ID value - assigned/set by the API Gateway based on the session
    schema:
      $ref: '#/definitions/error-response'
...
definitions:
...
  error-response:
    type: object
    x-nullable: false
    title: Error Response
    description: Standard error format
    properties:
      Code:
        description: The error code
        example: "404"
        type: string
      Message:
        description: The error message
        example: "unable to find project Acme, Inc."
        type: string
      x-request-id:
        description: The unique request ID value - assigned/set by the API Gateway or the API based on the login session
        example: 'b1e86e26-d8c8-4fd8-9f8d-5c723d5dac9f'
        type: string
  success-response:
    type: object
    x-nullable: false
    title: Success Response
    description: Standard success format
    properties:
      Code:
        description: The success code
        example: "202"
        type: string
      Message:
        description: The success message
        example: "unable to find project Acme, Inc."
        type: string
      x-request-id:
        description: The unique request ID value - assigned/set by the API Gateway or the API based on the login session
        example: 'b1e86e26-d8c8-4fd8-9f8d-5c723d5dac9f'
        type: string